Fixed some tests.| #1327
