+++ FlowCrypt/src/androidTest/java/com/flowcrypt/email/ui/activity/PrivateKeysListFragmentTest.kt
@@ -8,22 +8,13 @@ package com.flowcrypt.email.ui.activity
 import android.app.Activity
 import android.app.Instrumentation
 import android.content.ComponentName
-import android.content.Intent
-import android.os.Environment
-import android.text.TextUtils
-import android.text.format.DateFormat
-import androidx.core.content.FileProvider
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.action.ViewActions.scrollTo
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.contrib.RecyclerViewActions
 import androidx.test.espresso.intent.Intents.intending
-import androidx.test.espresso.intent.matcher.IntentMatchers.hasAction
-import androidx.test.espresso.intent.matcher.IntentMatchers.hasCategories
 import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
-import androidx.test.espresso.intent.matcher.IntentMatchers.hasType
 import androidx.test.espresso.matcher.ViewMatchers.hasTextColor
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
@@ -31,12 +22,9 @@ import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.rules.activityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.MediumTest
-import androidx.test.internal.runner.junit4.statement.UiThreadStatement
-import com.flowcrypt.email.Constants
 import com.flowcrypt.email.R
 import com.flowcrypt.email.TestConstants
 import com.flowcrypt.email.base.BaseTest
-import com.flowcrypt.email.junit.annotations.NotReadyForCI
 import com.flowcrypt.email.matchers.CustomMatchers.Companion.withEmptyRecyclerView
 import com.flowcrypt.email.matchers.CustomMatchers.Companion.withRecyclerViewItemCount
 import com.flowcrypt.email.model.KeyImportDetails
@@ -46,22 +34,15 @@ import com.flowcrypt.email.rules.ClearAppSettingsRule
 import com.flowcrypt.email.rules.RetryRule
 import com.flowcrypt.email.rules.ScreenshotTestRule
 import com.flowcrypt.email.ui.activity.settings.SettingsActivity
-import com.flowcrypt.email.util.GeneralUtil
 import com.flowcrypt.email.util.PrivateKeysManager
 import com.flowcrypt.email.util.TestGeneralUtil
-import org.hamcrest.CoreMatchers.hasItem
-import org.hamcrest.Matchers.equalTo
 import org.hamcrest.Matchers.not
-import org.hamcrest.core.AllOf.allOf
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
 import org.junit.rules.TestRule
 import org.junit.runner.RunWith
-import java.io.File
-import java.util.ArrayList
-import java.util.Date
 
 /**
  * @author Denis Bondarenko
@@ -131,84 +112,8 @@ class PrivateKeysListFragmentTest : BaseTest() {
 
   @Test
   fun testShowKeyDetailsScreen() {
-    selectFirstKey()
-  }
-
-  @Test
-  fun testKeyDetailsShowPubKey() {
-    selectFirstKey()
-    val keyDetails = addPrivateKeyToDatabaseRule.pgpKeyDetails
-    onView(withId(R.id.btnShowPubKey))
-      .check(matches(isDisplayed()))
-      .perform(click())
-    onView(withText(keyDetails.publicKey))
-  }
-
-  @Test
-  fun testKeyDetailsCopyToClipBoard() {
-    selectFirstKey()
-    val details = addPrivateKeyToDatabaseRule.pgpKeyDetails
-    onView(withId(R.id.btnCopyToClipboard))
-      .check(matches(isDisplayed()))
-      .perform(click())
-    isToastDisplayed(getResString(R.string.copied))
-    UiThreadStatement.runOnUiThread { checkClipboardText(details.publicKey) }
-  }
-
-  @Test
-  @NotReadyForCI
-  fun testKeyDetailsShowPrivateKey() {
-    selectFirstKey()
-    onView(withId(R.id.btnShowPrKey))
-      .perform(scrollTo())
-      .perform(click())
-    isToastDisplayed(getResString(R.string.see_backups_to_save_your_private_keys))
-  }
-
-  @Test
-  fun testKeyDetailsCheckDetails() {
-    selectFirstKey()
-    val details = addPrivateKeyToDatabaseRule.pgpKeyDetails
-
-    onView(withId(R.id.textViewFingerprint))
-      .check(
-        matches(
-          withText(
-            getHtmlString(
-              getResString(
-                R.string.template_fingerprint,
-                GeneralUtil.doSectionsInText(" ", details.fingerprint, 4)!!
-              )
-            )
-          )
-        )
-      )
-
-    onView(withId(R.id.textViewDate))
-      .check(
-        matches(
-          withText(
-            getHtmlString(
-              getResString(
-                R.string.template_date,
-                DateFormat.getMediumDateFormat(getTargetContext()).format(Date(details.created))
-              )
-            )
-          )
-        )
-      )
-
-    onView(withId(R.id.tVPassPhraseVerification))
-      .check(matches(withText(getResString(R.string.stored_pass_phrase_matched))))
-
-    val emails = ArrayList<String>()
-
-    for (pgpContact in details.pgpContacts) {
-      emails.add(pgpContact.email)
-    }
-
-    onView(withId(R.id.tVUsers))
-      .check(matches(withText(getResString(R.string.template_users, TextUtils.join(", ", emails)))))
+    onView(withId(R.id.recyclerViewKeys))
+      .perform(RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(0, click()))
   }
 
   @Test
@@ -229,44 +134,5 @@ class PrivateKeysListFragmentTest : BaseTest() {
     onView(withId(R.id.tVPassPhraseVerification))
       .check(matches(withText(getResString(R.string.stored_pass_phrase_mismatch))))
       .check(matches(hasTextColor(R.color.red)))
-  }
-
-  @Test
-  fun testKeyDetailsSavePubKeyToFileWhenFileIsNotExist() {
-    selectFirstKey()
-    val details = addPrivateKeyToDatabaseRule.pgpKeyDetails
-
-    val file =
-      File(
-        getTargetContext().getExternalFilesDir(Environment.DIRECTORY_DOCUMENTS),
-        "0x" + details.fingerprint + ".asc"
-      )
-
-    if (file.exists()) {
-      file.delete()
-    }
-
-    val resultData = Intent()
-    resultData.data =
-      FileProvider.getUriForFile(getTargetContext(), Constants.FILE_PROVIDER_AUTHORITY, file)
-
-    intending(
-      allOf(
-        hasAction(Intent.ACTION_CREATE_DOCUMENT),
-        hasCategories(hasItem(equalTo(Intent.CATEGORY_OPENABLE))),
-        hasType(Constants.MIME_TYPE_PGP_KEY)
-      )
-    )
-      .respondWith(Instrumentation.ActivityResult(Activity.RESULT_OK, resultData))
-
-    onView(withId(R.id.btnSaveToFile))
-      .check(matches(isDisplayed()))
-      .perform(click())
-    isToastDisplayed(getResString(R.string.saved))
-  }
-
-  private fun selectFirstKey() {
-    onView(withId(R.id.recyclerViewKeys))
-      .perform(RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(0, click()))
   }
 }
