+++ FlowCrypt/src/androidTest/resources/base_account_settings.json
@@ -1,22 +1,22 @@
 {
   "accountType": "flowcrypt.test",
-  "areContactsLoaded": true,
+  "contactsLoaded": true,
   "email": "default@flowcrypt.test",
-  "imapIsUseSslTls": true,
-  "imapIsUseStarttls": false,
+  "imapUseSslTls": true,
+  "imapUseStarttls": false,
   "imapPort": 993,
   "imapServer": "10.0.2.2",
   "isActive": true,
   "isEnabled": true,
   "isRestoreAccessRequired": false,
-  "isShowOnlyEncrypted": false,
+  "showOnlyEncrypted": false,
   "password": "qwerty1234",
-  "smtpIsUseSslTls": false,
-  "smtpIsUseStarttls": true,
+  "smtpUseSslTls": false,
+  "smtpUseStarttls": true,
   "smtpPassword": "qwerty1234",
   "smtpPort": 587,
   "smtpServer": "10.0.2.2",
   "smtpUsername": "default@flowcrypt.test",
-  "useCustomSignForSmtp": true,
+  "smtpUseCustomSign": true,
   "username": "default@flowcrypt.test"
 }
