+++ FlowCrypt/src/androidTest/resources/no.pub@org-rules-test.flowcrypt.com.json
@@ -7,10 +7,14 @@
   "photoUrl": "https://lh3.googleusercontent.com/-KqfTq1IKkhw/AAAAAAAAAAI/AAAAAAAAAAA/ACLGyWBZ5bzZtjs2GmmC1ArEskmiLFQWQQ/mo/photo.jpg?sz=128",
   "imapPort": 993,
   "imapServer": "imap.flowcrypt.com",
-  "useCustomSignForSmtp": false,
+  "smtpUseCustomSign": false,
   "password": "qwerty1234",
   "smtpPort": 25,
   "smtpServer": "smtp.flowcrypt.com",
   "username": "no.pub@org-rules-test.flowcrypt.com",
-  "domainRules": "ENFORCE_ATTESTER_SUBMIT"
+  "clientConfiguration": {
+    "flags": [
+      "ENFORCE_ATTESTER_SUBMIT"
+    ]
+  }
 }
